nginx_test/was-7-flask/requirements.txt
# Flask and WSGI
Flask==3.0.0
gunicorn==21.2.0
gevent==23.9.1

# Redis
redis==5.0.1
redis-py-cluster==2.1.3

# Kafka
confluent-kafka==2.3.0

# Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-requests==0.43b0

# Logging
python-json-logger==2.0.7

# Utilities
requests==2.31.0
python-dotenv==1.0.0
